ui: More cleanups and refactoring
