+++ ui/src/app/api/index.js
@@ -1,0 +1,8 @@
+
+import PlaygroundAPI from './endpoint';
+
+export {
+    PlaygroundAPIImages, PlaygroundAPINetworks,
+    PlaygroundAPIServices, PlaygroundAPIProxies} from './handler';
+
+export default PlaygroundAPI;
